requirements.txt
openai-whisper==20231117
ffmpeg-python==0.2.0
pyannote.audio==3.1.1
torch==2.1.2
torchaudio==2.1.2
tqdm==4.66.1
numpy>=1.26.0
python-dotenv==1.0.1
requests==2.31.0 